ehhh ?

+++ wServer/logic/BehaviorDb.cs
@@ -106,10 +106,10 @@ namespace wServer.logic
                 {
                     var loot = new Loot(defs);
                     rootState.Death += (sender, e) => loot.Handle((Enemy)e.Host, e.Time);
-                    InitDb.Definitions.Add((ushort)dat.IdToObjectType[objType], new Tuple<State, Loot>(rootState, loot));
+                    InitDb.Definitions.Add(dat.IdToObjectType[objType], new Tuple<State, Loot>(rootState, loot));
                 }
                 else
-                    InitDb.Definitions.Add((ushort)dat.IdToObjectType[objType], new Tuple<State, Loot>(rootState, null));
+                    InitDb.Definitions.Add(dat.IdToObjectType[objType], new Tuple<State, Loot>(rootState, null));
                 return this;
             }
         }
